PHP 8 tweaks + new character to make lists from: '○'

+++ lib/PreformattedOutput.php
@@ -152,7 +152,7 @@ class PreformattedOutput
         } elseif ($request['request'] === 'RS') {
             array_shift($lines);
             if (count($request['arguments'])) {
-                self::$addIndent = (int)round(Unit::normalize($request['arguments'][0], 'n', 'm'));
+                self::$addIndent = (int)round((float)Unit::normalize($request['arguments'][0], 'n', 'm'));
             } else {
                 self::$addIndent = 4;
             }
